[fix] #162 - 공연 삭제 로직 변경 (#163)
* [#162] refactor(Performance): Promotion 엔티티 양방향 매핑

* [#162] refactor(BookingRepository): JPQL을 사용해 회차 아이디가 담긴 리스트의 회차 아이디가 예매 테이블에도 존재하는지 비교하도록 변경

* [#162] refactor(PerformanceController): 메서드 네이밍 리팩토링

* [#162] refactor(PerformanceErrorCode): 에러 상태 변경 및 해당 공연의 메이커가 아닐 때 발생하는 에러메시지 추가

* [#162] refactor(PerformanceManagementService): 변경된 JPA 메서드에 맞추어 서비스 로직 변경 및 해당 공연의 메이커만 공연삭제가 가능하도록 변경

* [#162] refactor(PerformanceService): 변경된 JPA 메서드에 맞추어 공연 수정 시 공연 정보를 반환하는 서비스 로직 변경

* [#162] fix(TicketController): 예매자 삭제 시 불필요한 pathVariable 속성 삭제

* [#162] refactor(ScheduleRepository): 공연 id에 해당하는 모든 회차 id를 리스트로 반환하는 메서드 추가

* [#162] refactor(TicketService): 해당 공연의 메이커가 아니면 예매자를 삭제하지 못하는 로직 추가

+++ src/main/java/com/beat/domain/performance/application/PerformanceManagementService.java
@@ -18,10 +18,8 @@ import com.beat.domain.schedule.dao.ScheduleRepository;
 import com.beat.domain.schedule.domain.Schedule;
 import com.beat.domain.staff.dao.StaffRepository;
 import com.beat.domain.staff.domain.Staff;
-import com.beat.domain.user.dao.UserRepository;
 import com.beat.domain.user.domain.Users;
-import com.beat.domain.user.exception.UserErrorCode;
-import com.beat.global.common.exception.BadRequestException;
+import com.beat.global.common.exception.ForbiddenException;
 import com.beat.global.common.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -164,14 +162,24 @@ public class PerformanceManagementService {
 
     @Transactional
     public void deletePerformance(Long memberId, Long performanceId) {
-        memberRepository.findById(memberId).orElseThrow(() -> new NotFoundException(UserErrorCode.USER_NOT_FOUND));
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new NotFoundException(MemberErrorCode.MEMBER_NOT_FOUND));
+
+        Long userId = member.getUser().getId();
 
         Performance performance = performanceRepository.findById(performanceId)
                 .orElseThrow(() -> new NotFoundException(PerformanceErrorCode.PERFORMANCE_NOT_FOUND));
 
-        boolean hasBookings = bookingRepository.existsBySchedulePerformanceId(performanceId);
+        if (!performance.getUsers().getId().equals(userId)) {
+            throw new ForbiddenException(PerformanceErrorCode.NOT_PERFORMANCE_OWNER);
+        }
+
+        List<Long> scheduleIds = scheduleRepository.findIdsByPerformanceId(performanceId);
+
+        boolean hasBookings = bookingRepository.existsByScheduleIdIn(scheduleIds);
+
         if (hasBookings) {
-            throw new BadRequestException(PerformanceErrorCode.PERFORMANCE_DELETE_FAILED);
+            throw new ForbiddenException(PerformanceErrorCode.PERFORMANCE_DELETE_FAILED);
         }
 
         performanceRepository.delete(performance);
